+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/AbstractFetchDownloadService.kt
@@ -111,6 +111,7 @@ abstract class AbstractFetchDownloadService : Service() {
     internal var downloadJobs = mutableMapOf<String, DownloadJobState>()
 
     protected abstract val fileSizeFormatter: FileSizeFormatter
+    protected abstract val dateTimeProvider: DateTimeProvider
 
     // TODO Move this to browser store and make immutable:
     // https://github.com/mozilla-mobile/android-components/issues/7050
@@ -124,6 +125,7 @@ abstract class AbstractFetchDownloadService : Service() {
         var notifiedStopped: Boolean = false,
         var lastNotificationUpdate: Long = 0L,
         var createdTime: Long = System.currentTimeMillis(),
+        val downloadEstimator: DownloadEstimator? = null,
     ) {
         internal fun canUpdateNotification(): Boolean {
             return isUnderNotificationUpdateLimit() && !notifiedStopped
@@ -303,6 +305,14 @@ abstract class AbstractFetchDownloadService : Service() {
             state = download.copy(status = actualStatus, notificationId = foregroundServiceId),
             foregroundServiceId = foregroundServiceId,
             status = actualStatus,
+            downloadEstimator = if (download.contentLength == null) {
+                null
+            } else {
+                DownloadEstimator(
+                    totalBytes = download.contentLength!!,
+                    dateTimeProvider = dateTimeProvider,
+                )
+            },
         )
 
         store.dispatch(DownloadAction.UpdateDownloadAction(downloadJobState.state))