+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/downloads/DownloadService.kt
@@ -6,6 +6,7 @@ package org.mozilla.samples.browser.downloads
 
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
+import mozilla.components.feature.downloads.DateTimeProvider
 import mozilla.components.feature.downloads.FileSizeFormatter
 import mozilla.components.support.base.android.NotificationsDelegate
 import org.mozilla.samples.browser.ext.components
@@ -15,4 +16,5 @@ class DownloadService : AbstractFetchDownloadService() {
     override val store: BrowserStore by lazy { components.store }
     override val notificationsDelegate: NotificationsDelegate by lazy { components.notificationsDelegate }
     override val fileSizeFormatter: FileSizeFormatter by lazy { components.fileSizeFormatter }
+    override val dateTimeProvider: DateTimeProvider by lazy { components.dateTimeProvider }
 }